feat: implement dashboard session handling and add sign-out button
