+++ deno.json
@@ -5,8 +5,8 @@
     "cli": "echo \"import '\\$fresh/src/dev/cli.ts'\" | deno run --unstable-kv -A -",
     "manifest": "deno task cli manifest $(pwd)",
     "start": "deno run --unstable-kv --env -A --watch=src/static/,src/routes/ src/dev.ts",
-    "build": "deno run -A src/dev.ts build",
-    "preview": "deno run -A src/main.ts",
+    "build": "deno run --unstable-kv -A src/dev.ts build",
+    "preview": "deno run --unstable-kv -A src/main.ts",
     "update": "deno run -A -r https://fresh.deno.dev/update ."
   },
   "lint": { "rules": { "tags": ["fresh", "recommended"] } },
