+++ src/fresh.gen.ts
@@ -15,15 +15,9 @@ import * as $index from "./routes/index.tsx";
 import * as $login from "./routes/login.tsx";
 import * as $protected_dashboard from "./routes/protected/dashboard.tsx";
 import * as $carousel from "./islands/carousel.tsx";
-import * as $dashboard from "./islands/dashboard.tsx";
 import * as $teamcards from "./islands/teamcards.tsx";
+import * as $typewriter from "./islands/typewriter.tsx";
 import type {Manifest} from "$fresh/server.ts";
-import * as $
-
-typewrit
-er
-from
-"./islands/typewriter.tsx";
 
 const manifest = {
   routes: {
@@ -37,7 +31,6 @@ const manifest = {
   },
   islands: {
     "./islands/carousel.tsx": $carousel,
-    "./islands/dashboard.tsx": $dashboard,
     "./islands/teamcards.tsx": $teamcards,
     "./islands/typewriter.tsx": $typewriter,
   },
